yoda: Support for "pile"-d options
If an option is specified several times, then yoda will
report (in yopts structure) only the last value given.

Options with an attribute "pile" (it can be "set" and "unset")
collect all occurences of on option in the array. The number
of elements is given in the "<option>_nr" variable on yopts
structure.

Signed-off-by: Pavel Emelyanov <xemul@sacred.ru>