yoda: Parse arguments in-order, rather than at the end of getopt loop
There are some scenarios (vzctl) when we need to keep track of arguments
in the argv[] array, rather than letting them all permutate to the end
of it and handle them later.

Now yoda generates parser, that scans argv-s one by one, assigning args
(if any) in-order, and after assignment of the last declared arg -- switch
the getopt_long to permutating mode.

Signed-off-by: Pavel Emelyanov <xemul@sacred.ru>