yoda: Classify options for opt vectors switch
We'll be in situations when we'd have to scan for more than one
options array depending on current parse status.

The plan is to switch arrays based on 'req_for'-s and 'for'-s.
This patch generates options groups. Each group matches some
expression (currently -- only a single argument match) and
is about to switch options array.

Signed-off-by: Pavel Emelyanov <xemul@sacred.ru>