yoda: Swich options array
All options are groupped based on their for and req_for keywords.
Based on these groups option arrays (sets) are formed. When an
action on which options depend happens, the currently scanned array
is switched to new one and new set of options apply.

Options that do not have for/req_for fields are parsed in every
array.

Signed-off-by: Pavel Emelyanov <xemul@sacred.ru>